Update: H.C.Dependencies を internal 式へ
